+++ torch_geometric/sampler/base.py
@@ -1,5 +1,6 @@
 import copy
 import math
+import warnings
 from abc import ABC
 from dataclasses import dataclass
 from enum import Enum
@@ -9,6 +10,7 @@ import torch
 from torch import Tensor
 
 from torch_geometric.data import Data, FeatureStore, GraphStore, HeteroData
+from torch_geometric.sampler.utils import to_bidirectional
 from torch_geometric.typing import EdgeType, EdgeTypeStr, NodeType, OptTensor
 from torch_geometric.utils.mixin import CastMixin
 
@@ -33,6 +35,13 @@ class DataType(Enum):
         raise ValueError(f"Expected a 'Data', 'HeteroData', or a tuple of "
                          f"'FeatureStore' and 'GraphStore' "
                          f"(got '{type(data)}')")
+
+
+class GraphMode(Enum):
+    r"""The graph mode of the returned subgraph."""
+    directional = 'directional'
+    bidirectional = 'bidirectional'
+    induced_subgraph = 'induced_subgraph'
 
 
 @dataclass
@@ -144,6 +153,23 @@ class SamplerOutput(CastMixin):
     # API for the expected output of a sampler.
     metadata: Optional[Any] = None
 
+    def to_bidirectional(self) -> 'SamplerOutput':
+        r"""Converts the sampled subgraph into a bidirectional variant, in
+        which all sampled edges are guaranteed to be bidirectional."""
+        out = copy.copy(self)
+
+        out.row, out.col, out.edge = to_bidirectional(
+            row=self.row,
+            col=self.col,
+            rev_row=self.row,
+            rev_col=self.col,
+            edge_id=self.edge,
+            rev_edge_id=self.edge,
+        )
+        out.num_sampled_edges = None
+
+        return out
+
 
 @dataclass
 class HeteroSamplerOutput(CastMixin):
@@ -190,6 +216,45 @@ class HeteroSamplerOutput(CastMixin):
     # TODO(manan): refine this further; it does not currently define a proper
     # API for the expected output of a sampler.
     metadata: Optional[Any] = None
+
+    def to_bidirectional(self) -> 'SamplerOutput':
+        r"""Converts the sampled subgraph into a bidirectional variant, in
+        which all sampled edges are guaranteed to be bidirectional."""
+        out = copy.copy(self)
+        out.row = copy.copy(self.row)
+        out.col = copy.copy(self.col)
+
+        edge_types = self.row.keys()
+        edge_types = [k for k in edge_types if not k[1].startswith('rev_')]
+        for edge_type in edge_types:
+            src, rel, dst = edge_type
+            rev_edge_type = (dst, f'rev_{rel}', src)
+
+            if src == dst and rev_edge_type not in self.row:
+                out.row[edge_type], out.col[edge_type], _ = to_bidirectional(
+                    row=self.row[edge_type],
+                    col=self.col[edge_type],
+                    rev_row=self.row[edge_type],
+                    rev_col=self.col[edge_type],
+                )
+            elif rev_edge_type in self.row:
+                out.row[edge_type], out.col[edge_type], _ = to_bidirectional(
+                    row=self.row[edge_type],
+                    col=self.col[edge_type],
+                    rev_row=self.row[rev_edge_type],
+                    rev_col=self.col[rev_edge_type],
+                )
+                out.row[rev_edge_type] = out.col[edge_type]
+                out.col[rev_edge_type] = out.row[edge_type]
+            else:
+                warnings.warn(f"Cannot convert to bidirectional graph since "
+                              f"the edge type {edge_type} does not seem to "
+                              f"have a reverse edge type")
+
+        out.edge = None
+        out.num_sampled_edges = None
+
+        return out
 
 
 @dataclass(frozen=True)